requirements.txt
fastapi
uvicorn
pymongo
google-generativeai
pdfminer.six
spacy
python-dotenv
